+++ src/main/java/racingcar/validator/CarsValidator.java
@@ -1,4 +1,4 @@
-package racingcar.util;
+package racingcar.validator;
 
 import racingcar.domain.Car;
